Refacto: Nodes & Edges managing & AddTable in specific hooks
